Break down the reconciliation routine into smaller sub functions, move the routine from Setup to eclair package. Remove startup spec (replaced by IntegrationSpec)
